cheviz: First attempt to get a classic dataframe out of a chess game
Had to slightly change the API around, and it feels a bit like cheating
because now the code in the previous notebooks didn't grow exactly the
way the text around it would suggest.

Anyhow, we give the board to the engine, get a score and put it together
with our own metrics into a Pandas dataframe.